Consider balance_change assertions to always be at the beginning of the day
This matches semantics of the built-in balance assertions
([beginning-vs-end-of-day](https://beancount.github.io/docs/balance_assertions_in_beancount.html#beginning-vs-end-of-day))

NOTE: While this better matches the built-in balance assertions, this is
a change in current behavior. Users that intentionally or
unintentionally rely on entry ordering might be broken by this.